// src/server/types.ts
export interface ScoredShow {
    score: number;
    show: Show;
}

export interface Show {
    id: number;
    url: string;
    name: string;
    type: string;
    language: string;
    genres: string[];
    status: string;
    runtime: number;
    averageRuntime: number;
    premiered: string;
    ended: string;
    officialSite: string;
    schedule: {
        time: string;
        days: string[];
    }
    rating: {
        average?: number;
    }
    weight: number;
    network: {
        id: number;
        name: string;
        country: {
            name: string;
            code: string;
            timezone: string;
        }
        officialSite: string;
    }
    webChannel: unknown;
    dvdCountry: unknown;
    externals: {
        tvrage?: number;
        thetvdb: number;
        imdb: string;
    }
    image?: {
        medium: string;
        original: string;
    }
    summary: string;
    updated: number;
    _links: {
        self: {
            href: string;
        }
        previousepisode: {
            href: string;
            name: string;
        }
    }
}

export interface Episode {
    id: number;
    url: string;
    name: string;
    season: number;
    number?: number;
    type: string;
    airdate: string;
    airtime: string;
    airstamp: string;
    runtime: number;
    rating: {
        average: number;
    }
    image?: {
        medium: string;
        original: string;
    }
    summary: string;
    _links: {
        self: {
            href: string;
        }
        show: {
            href: string;
            name: string;
        }
    }
}